More campaign options reshuffling
