+++ MekHQ/resources/mekhq/resources/CampaignOptionsDialog.properties
@@ -359,6 +359,10 @@ lblNoDeliveriesInTransit.text=No Deliveries in Transit
 lblNoDeliveriesInTransit.tooltip=When this campaign option is enabled, parts and units will not be delivered unless \
   the campaign is planetside. Delivery time will still go down, while in transit, but will stop at 1 day. This means \
   delivery will occur the day after you arrive at your destination.
+# Personnel > Equipment page
+lblPersonnelEquipmentPage.text=Equipment
+lblEquipmentGeneralPanel.text=General
+lblEquipmentGeneralPanel.summary=Quartermaster defaults that apply to every kind of equipment kit.
 lblArmorKitsPanel.text=Armor Kits
 lblArmorKitsPanel.summary=Quartermaster defaults for issuing personal armor kits.
 armorKitNone.text=None
@@ -403,11 +407,11 @@ kitTooltip.PocketTranscriber=+2 to Negotiation checks.
 kitTooltip.Telescan=+1 to Administration checks.
 lblToolKitsPanel.text=Tool Kits
 lblToolKitsPanel.summary=Quartermaster defaults for issuing tool kits, by support profession.
-lblMekTechDefaultToolKit.text=Mek Tech Default Kit
-lblMekTechDefaultToolKit.tooltip=The tool kit newly recruited Mek Techs try to equip on joining, drawn from local stores.
+lblMekTechDefaultToolKit.text=MekTech Default Kit
+lblMekTechDefaultToolKit.tooltip=The tool kit newly recruited MekTechs try to equip on joining, drawn from local stores.
 lblMechanicDefaultToolKit.text=Mechanic Default Kit
 lblMechanicDefaultToolKit.tooltip=The tool kit newly recruited Mechanics try to equip on joining, drawn from local stores.
-lblAeroTechDefaultToolKit.text=Aero Tech Default Kit
+lblAeroTechDefaultToolKit.text=AeroTek Default Kit
 lblAeroTechDefaultToolKit.tooltip=The tool kit newly recruited Aerospace Techs try to equip on joining, drawn from local stores.
 lblBATechDefaultToolKit.text=Battle Armor Tech Default Kit
 lblBATechDefaultToolKit.tooltip=The tool kit newly recruited Battle Armor Techs try to equip on joining, drawn from local stores.
@@ -526,6 +530,9 @@ personnelGeneralPage.title=General
 personnelGeneralPage.border="The merc's worst enemy is usually the employer's High Command."\
   <br><i>Colonel Jaime Wolf\
   <br>Commander of Wolf's Dragoons</i>
+personnelEquipmentPage.title=Equipment
+personnelEquipmentPage.border="An army marches on its stomach and dies without its quartermaster."\
+  <br><i>Attributed to a Star League Quartermaster Corps training manual</i>
 personnelInformationPage.title=Personnel Information
 personnelInformationPage.border="Why continue a wasteful campaign from an untenable position?"\
   <br><i>From a letter written by Takashi Kurita to his son Theodore, Prince of Luthien\
